rag/requirements.txt
# SochDB RAG System - Requirements

# Core dependencies
sochdb>=0.4.0
python-dotenv>=1.0.0
numpy>=1.20.0

# Azure OpenAI
openai>=1.0.0

# PDF processing
PyMuPDF>=1.23.0

# Testing
pytest>=7.0.0

# Optional: Local embeddings fallback
# sentence-transformers>=2.0.0
